backend/scripts/test_end_to_end_workflow.py
"""
End-to-End Workflow Test for Scoring Strategy Refactoring

Tests the complete workflow:
1. Enter a score via API
2. Verify calculated values are stored in database
3. Verify leaderboard reflects the new score without recalculation

This test proves the "calculate once, store, display many" architecture works end-to-end.

Usage:
    python scripts/test_end_to_end_workflow.py
"""

import sys
from pathlib import Path
import requests

# Add parent directory to path for imports
sys.path.insert(0, str(Path(__file__).parent.parent))

from sqlmodel import Session, select
from core.database import engine
from models.scorecard import Scorecard
from models.participant import Participant
from models.event import Event

# API base URL
BASE_URL = "http://localhost:8000/api/v1"


class EndToEndWorkflowTester:
    """Tests end-to-end scoring workflow through API"""

    def __init__(self):
        self.token = None
        self.session = Session(engine)
        self.test_results = {
            'total_tests': 0,
            'passed': 0,
            'failed': 0,
            'errors': []
        }

    def run_all_tests(self):
        """Run all workflow tests"""
        print("\n" + "="*60)
        print("  End-to-End Workflow Test")
        print("="*60 + "\n")

        try:
            # Step 1: Login to get auth token
            if not self.login():
                print("[ERROR] Failed to login - cannot proceed with tests")
                return

            # Step 2: Test score entry workflow
            self.test_score_entry_workflow()

            # Print summary
            self.print_summary()

        except Exception as e:
            print(f"\n[ERROR] Test suite failed: {e}")
            import traceback
            traceback.print_exc()
        finally:
            self.session.close()

    def login(self):
        """Login to get authentication token"""
        print("[SETUP] Logging in to API...")

        try:
            response = requests.post(
                f"{BASE_URL}/auth/login",
                json={
                    "email": "admin@abhimatagolf.com",
                    "password": "admin123"
                }
            )

            if response.status_code == 200:
                self.token = response.json()["access_token"]
                print("[OK] Login successful\n")
                return True
            else:
                print(f"[ERROR] Login failed: {response.status_code} - {response.text}")
                return False

        except Exception as e:
            print(f"[ERROR] Login exception: {e}")
            return False

    def test_score_entry_workflow(self):
        """Test complete score entry to leaderboard workflow"""
        print("[TEST] Score Entry to Leaderboard Workflow")
        print("-" * 60)

        # Get an event and participant to test with
        event = self.session.exec(select(Event)).first()
        if not event:
            print("  [SKIP] No events in database")
            return

        participant = self.session.exec(
            select(Participant).where(Participant.event_id == event.id)
        ).first()
        if not participant:
            print(f"  [SKIP] No participants for event {event.name}")
            return

        # Test with hole 5 (updating existing score)
        test_hole_number = 5
        test_strokes = 6  # New score to enter

        # Step 1: Enter score via API
        self.test_results['total_tests'] += 1
        print(f"\n  Step 1: Enter score via API")
        print(f"          Participant: {participant.name}")
        print(f"          Event: {event.name}")
        print(f"          Hole Number: {test_hole_number}")
        print(f"          Strokes: {test_strokes}")

        try:
            headers = {"Authorization": f"Bearer {self.token}"}
            # API uses query parameters: POST /scorecards/?participant_id=X&hole_number=Y&strokes=Z
            response = requests.post(
                f"{BASE_URL}/scorecards/",
                params={
                    "participant_id": participant.id,
                    "hole_number": test_hole_number,
                    "strokes": test_strokes
                },
                headers=headers
            )

            if response.status_code in [200, 201]:
                print(f"  [PASS] Score entered successfully")
                scorecard_data = response.json()
                print(f"          Response: Strokes={scorecard_data['strokes']}, "
                      f"Net Score={scorecard_data.get('net_score')}, "
                      f"Points={scorecard_data.get('points')}")
            else:
                print(f"  [FAIL] Score entry failed: {response.status_code}")
                print(f"          Error: {response.text}")
                self.test_results['failed'] += 1
                self.test_results['errors'].append(f"Score entry API failed: {response.status_code}")
                return

        except Exception as e:
            print(f"  [FAIL] Score entry exception: {e}")
            self.test_results['failed'] += 1
            self.test_results['errors'].append(f"Score entry exception: {e}")
            return

        # Step 2: Verify calculated values in database
        print(f"\n  Step 2: Verify values stored in database")

        # Refresh session to get latest data
        self.session.expire_all()

        # Get hole_id from hole_number
        from models.course import Hole
        hole = self.session.exec(
            select(Hole)
            .where(Hole.course_id == event.course_id)
            .where(Hole.number == test_hole_number)
        ).first()

        if not hole:
            print(f"  [FAIL] Hole {test_hole_number} not found")
            self.test_results['failed'] += 1
            return

        scorecard = self.session.exec(
            select(Scorecard)
            .where(Scorecard.participant_id == participant.id)
            .where(Scorecard.hole_id == hole.id)
        ).first()

        if not scorecard:
            print(f"  [FAIL] Scorecard not found in database")
            self.test_results['failed'] += 1
            self.test_results['errors'].append("Scorecard not in database")
            return

        print(f"          Strokes: {scorecard.strokes}")
        print(f"          Net Score: {scorecard.net_score}")
        print(f"          Points: {scorecard.points}")

        # Verify values are calculated and stored
        if scorecard.strokes == test_strokes:
            print(f"  [PASS] Strokes stored correctly")
        else:
            print(f"  [FAIL] Strokes mismatch: expected {test_strokes}, got {scorecard.strokes}")
            self.test_results['failed'] += 1
            return

        # For System36, verify points are calculated (should be > 0 for par or better)
        if event.scoring_type == "SYSTEM_36":
            if scorecard.points is not None and scorecard.points >= 0:
                print(f"  [PASS] Points calculated and stored: {scorecard.points}")
            else:
                print(f"  [FAIL] Points not calculated")
                self.test_results['failed'] += 1
                return

        # Verify net_score is calculated
        if scorecard.net_score is not None:
            print(f"  [PASS] Net score calculated and stored: {scorecard.net_score}")
        else:
            print(f"  [FAIL] Net score not calculated")
            self.test_results['failed'] += 1
            return

        # Step 3: Get leaderboard and verify it reflects the score
        print(f"\n  Step 3: Verify leaderboard reflects stored values")

        try:
            headers = {"Authorization": f"Bearer {self.token}"}
            response = requests.get(
                f"{BASE_URL}/leaderboards/event/{event.id}?use_cache=false",
                headers=headers
            )

            if response.status_code == 200:
                leaderboard = response.json()
                print(f"  [PASS] Leaderboard retrieved successfully")
                print(f"          Total participants: {leaderboard['total_participants']}")
                print(f"          Participants with scores: {leaderboard['participants_with_scores']}")

                # Find our participant in the leaderboard
                our_entry = None
                for entry in leaderboard['entries']:
                    if entry['participant_id'] == participant.id:
                        our_entry = entry
                        break

                if our_entry:
                    print(f"\n  [PASS] Participant found in leaderboard")
                    print(f"          Name: {our_entry['participant_name']}")
                    print(f"          Rank: {our_entry['rank']}")
                    print(f"          Gross Score: {our_entry['gross_score']}")
                    print(f"          Net Score: {our_entry['net_score']}")
                    print(f"          Points: {our_entry['system36_points']}")
                    print(f"          Holes Completed: {our_entry['holes_completed']}")

                    # Verify leaderboard is reading stored values (not recalculating)
                    # The leaderboard entry should include our new score
                    if our_entry['holes_completed'] > 0:
                        print(f"\n  [PASS] Leaderboard includes participant scores")
                        print(f"  [SUCCESS] End-to-end workflow complete!")
                        self.test_results['passed'] += 1
                    else:
                        print(f"  [FAIL] Leaderboard shows 0 holes completed")
                        self.test_results['failed'] += 1
                else:
                    print(f"  [FAIL] Participant not found in leaderboard")
                    self.test_results['failed'] += 1
            else:
                print(f"  [FAIL] Leaderboard retrieval failed: {response.status_code}")
                print(f"          Error: {response.text}")
                self.test_results['failed'] += 1
                self.test_results['errors'].append(f"Leaderboard API failed: {response.status_code}")

        except Exception as e:
            print(f"  [FAIL] Leaderboard exception: {e}")
            self.test_results['failed'] += 1
            self.test_results['errors'].append(f"Leaderboard exception: {e}")

        print()

    def print_summary(self):
        """Print test summary"""
        print("="*60)
        print("  Test Summary")
        print("="*60 + "\n")

        print(f"Total Tests: {self.test_results['total_tests']}")
        print(f"[PASS] Passed: {self.test_results['passed']}")
        print(f"[FAIL] Failed: {self.test_results['failed']}")

        if self.test_results['errors']:
            print("\nErrors:")
            for error in self.test_results['errors']:
                print(f"  - {error}")

        print("\n" + "="*60)

        # Exit code
        if self.test_results['failed'] == 0 and self.test_results['passed'] > 0:
            print("\n[SUCCESS] End-to-end workflow test passed!")
            print("\nKey Achievement:")
            print("- Score calculated ONCE during entry")
            print("- Values stored in database (net_score, points)")
            print("- Leaderboard reads stored values (NO recalculation)")
            print("- Strategy Pattern working end-to-end")
            sys.exit(0)
        else:
            print("\n[FAILURE] Some tests failed")
            sys.exit(1)


def main():
    """Main entry point"""
    tester = EndToEndWorkflowTester()
    tester.run_all_tests()


if __name__ == "__main__":
    main()
