backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.14
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
pydantic-settings==2.1.0
email-validator==2.1.0
python-dotenv==1.0.0
pandas>=2.2.0
openpyxl==3.1.2
python-socketio==5.10.0
